crete auto_traits.rs
