+++ src/test.rs
@@ -112,6 +112,7 @@ fn solve_goal(program_text: &str, goals: Vec<(&str, SolverChoice, &str)>) {
     }
 }
 
+mod auto_traits;
 mod coherence_goals;
 mod coinduction;
 mod cycle;
@@ -121,96 +122,6 @@ mod negation;
 mod projection;
 mod unify;
 mod wf_goals;
-
-#[test]
-fn auto_trait_without_impls() {
-    test! {
-        program {
-            #[auto] trait Send { }
-
-            struct i32 { }
-
-            struct Useless<T> { }
-
-            struct Data<T> {
-                data: T
-            }
-        }
-
-        goal {
-            i32: Send
-        } yields {
-            "Unique"
-        }
-
-        // No fields so `Useless<T>` is `Send`.
-        goal {
-            forall<T> {
-                Useless<T>: Send
-            }
-        } yields {
-            "Unique"
-        }
-
-        goal {
-            forall<T> {
-                if (T: Send) {
-                    Data<T>: Send
-                }
-            }
-        } yields {
-            "Unique"
-        }
-    }
-}
-
-#[test]
-fn auto_trait_with_impls() {
-    test! {
-        program {
-            #[auto] trait Send { }
-
-            struct i32 { }
-            struct f32 { }
-            struct Vec<T> { }
-
-            impl<T> Send for Vec<T> where T: Send { }
-            impl !Send for i32 { }
-        }
-
-        goal {
-            i32: Send
-        } yields {
-            "No possible solution"
-        }
-
-        goal {
-            f32: Send
-        } yields {
-            "Unique"
-        }
-
-        goal {
-            Vec<i32>: Send
-        } yields {
-            "No possible solution"
-        }
-
-        goal {
-            Vec<f32>: Send
-        } yields {
-            "Unique"
-        }
-
-        goal {
-            forall<T> {
-                Vec<T>: Send
-            }
-        } yields {
-            "No possible solution"
-        }
-    }
-}
 
 #[test]
 fn partial_overlap_2() {
