+++ src/test/auto_traits.rs
@@ -1,0 +1,93 @@
+//! Tests targeting auto traits specifically
+
+use super::*;
+
+#[test]
+fn auto_trait_without_impls() {
+    test! {
+        program {
+            #[auto] trait Send { }
+
+            struct i32 { }
+
+            struct Useless<T> { }
+
+            struct Data<T> {
+                data: T
+            }
+        }
+
+        goal {
+            i32: Send
+        } yields {
+            "Unique"
+        }
+
+        // No fields so `Useless<T>` is `Send`.
+        goal {
+            forall<T> {
+                Useless<T>: Send
+            }
+        } yields {
+            "Unique"
+        }
+
+        goal {
+            forall<T> {
+                if (T: Send) {
+                    Data<T>: Send
+                }
+            }
+        } yields {
+            "Unique"
+        }
+    }
+}
+
+#[test]
+fn auto_trait_with_impls() {
+    test! {
+        program {
+            #[auto] trait Send { }
+
+            struct i32 { }
+            struct f32 { }
+            struct Vec<T> { }
+
+            impl<T> Send for Vec<T> where T: Send { }
+            impl !Send for i32 { }
+        }
+
+        goal {
+            i32: Send
+        } yields {
+            "No possible solution"
+        }
+
+        goal {
+            f32: Send
+        } yields {
+            "Unique"
+        }
+
+        goal {
+            Vec<i32>: Send
+        } yields {
+            "No possible solution"
+        }
+
+        goal {
+            Vec<f32>: Send
+        } yields {
+            "Unique"
+        }
+
+        goal {
+            forall<T> {
+                Vec<T>: Send
+            }
+        } yields {
+            "No possible solution"
+        }
+    }
+}
